Split VoxelGeneratorGraph into generator and VoxelGraphFunction
