+++ generators/graph/voxel_graph_function.h
@@ -1,0 +1,182 @@
+#ifndef VOXEL_GRAPH_FUNCTION_H
+#define VOXEL_GRAPH_FUNCTION_H
+
+#include "../../util/godot/binder.h"
+#include "../../util/godot/resource.h"
+#include "program_graph.h"
+
+namespace zylann::voxel {
+
+// Generic processing graph made of operation nodes.
+class VoxelGraphFunction : public Resource {
+	GDCLASS(VoxelGraphFunction, Resource)
+public:
+	static const char *SIGNAL_NODE_NAME_CHANGED;
+
+	// Node indexes within the DB.
+	// Don't use these in saved data,
+	// they can change depending on which features the module is compiled with.
+	enum NodeTypeID {
+		NODE_CONSTANT,
+		NODE_INPUT_X,
+		NODE_INPUT_Y,
+		NODE_INPUT_Z,
+		NODE_OUTPUT_SDF,
+		NODE_ADD,
+		NODE_SUBTRACT,
+		NODE_MULTIPLY,
+		NODE_DIVIDE,
+		NODE_SIN,
+		NODE_FLOOR,
+		NODE_ABS,
+		NODE_SQRT,
+		NODE_FRACT,
+		NODE_STEPIFY,
+		NODE_WRAP,
+		NODE_MIN,
+		NODE_MAX,
+		NODE_DISTANCE_2D,
+		NODE_DISTANCE_3D,
+		NODE_CLAMP,
+		NODE_CLAMP_C,
+		NODE_MIX,
+		NODE_REMAP,
+		NODE_SMOOTHSTEP,
+		NODE_CURVE,
+		NODE_SELECT,
+		NODE_NOISE_2D,
+		NODE_NOISE_3D,
+		NODE_IMAGE_2D,
+		NODE_SDF_PLANE,
+		NODE_SDF_BOX,
+		NODE_SDF_SPHERE,
+		NODE_SDF_TORUS,
+		NODE_SDF_PREVIEW, // For debugging
+		NODE_SDF_SPHERE_HEIGHTMAP,
+		NODE_SDF_SMOOTH_UNION,
+		NODE_SDF_SMOOTH_SUBTRACT,
+		NODE_NORMALIZE_3D,
+		NODE_FAST_NOISE_2D,
+		NODE_FAST_NOISE_3D,
+		NODE_FAST_NOISE_GRADIENT_2D,
+		NODE_FAST_NOISE_GRADIENT_3D,
+		NODE_OUTPUT_WEIGHT,
+		NODE_OUTPUT_TYPE,
+#ifdef VOXEL_ENABLE_FAST_NOISE_2
+		NODE_FAST_NOISE_2_2D,
+		NODE_FAST_NOISE_2_3D,
+#endif
+		NODE_OUTPUT_SINGLE_TEXTURE,
+		NODE_EXPRESSION,
+		NODE_POWI, // pow(x, constant positive integer)
+		NODE_POW, // pow(x, y)
+		NODE_INPUT_SDF,
+		NODE_COMMENT,
+
+		NODE_TYPE_COUNT
+	};
+
+	void clear();
+
+	// Graph edition API
+	// Important: functions editing the graph are NOT thread-safe.
+	// They are expected to be used by the main thread (editor or game logic).
+
+	uint32_t create_node(NodeTypeID type_id, Vector2 position, uint32_t id = ProgramGraph::NULL_ID);
+	void remove_node(uint32_t node_id);
+
+	// Checks if the specified connection can be created
+	bool can_connect(
+			uint32_t src_node_id, uint32_t src_port_index, uint32_t dst_node_id, uint32_t dst_port_index) const;
+
+	// Checks if the specified connection is valid (without considering existing connections)
+	bool is_valid_connection(
+			uint32_t src_node_id, uint32_t src_port_index, uint32_t dst_node_id, uint32_t dst_port_index) const;
+
+	void add_connection(uint32_t src_node_id, uint32_t src_port_index, uint32_t dst_node_id, uint32_t dst_port_index);
+	void remove_connection(
+			uint32_t src_node_id, uint32_t src_port_index, uint32_t dst_node_id, uint32_t dst_port_index);
+	void get_connections(std::vector<ProgramGraph::Connection> &connections) const;
+
+	// Finds which source port is connected to the given destination.
+	// Returns false if `dst` has no inbound connection.
+	bool try_get_connection_to(ProgramGraph::PortLocation dst, ProgramGraph::PortLocation &out_src) const;
+
+	bool has_node(uint32_t node_id) const;
+
+	void set_node_name(uint32_t node_id, StringName name);
+	StringName get_node_name(uint32_t node_id) const;
+	uint32_t find_node_by_name(StringName name) const;
+
+	Variant get_node_param(uint32_t node_id, uint32_t param_index) const;
+	void set_node_param(uint32_t node_id, uint32_t param_index, Variant value);
+
+	static bool get_expression_variables(std::string_view code, std::vector<std::string_view> &vars);
+	void get_expression_node_inputs(uint32_t node_id, std::vector<std::string> &out_names) const;
+	void set_expression_node_inputs(uint32_t node_id, PackedStringArray names);
+
+	Variant get_node_default_input(uint32_t node_id, uint32_t input_index) const;
+	void set_node_default_input(uint32_t node_id, uint32_t input_index, Variant value);
+
+	bool get_node_default_inputs_autoconnect(uint32_t node_id) const;
+	void set_node_default_inputs_autoconnect(uint32_t node_id, bool enabled);
+
+	Vector2 get_node_gui_position(uint32_t node_id) const;
+	void set_node_gui_position(uint32_t node_id, Vector2 pos);
+
+	Vector2 get_node_gui_size(uint32_t node_id) const;
+	void set_node_gui_size(uint32_t node_id, Vector2 size);
+
+	NodeTypeID get_node_type_id(uint32_t node_id) const;
+	PackedInt32Array get_node_ids() const;
+	uint32_t generate_node_id() {
+		return _graph.generate_node_id();
+	}
+
+	unsigned int get_nodes_count() const;
+
+	// Editor
+
+#ifdef TOOLS_ENABLED
+	void get_configuration_warnings(PackedStringArray &out_warnings) const;
+
+	// Gets a hash that attempts to only change if the output of the graph is different.
+	// This is computed from the editable graph data, not the compiled result.
+	uint64_t get_output_graph_hash() const;
+#endif
+
+	// Internal
+
+	const ProgramGraph &get_graph() const;
+	void find_dependencies(uint32_t node_id, std::vector<uint32_t> &out_dependencies) const;
+	Dictionary get_graph_as_variant_data() const;
+	bool load_graph_from_variant_data(Dictionary data);
+
+private:
+	void register_subresource(Resource &resource);
+	void unregister_subresource(Resource &resource);
+	void register_subresources();
+	void unregister_subresources();
+	void _on_subresource_changed();
+
+	int _b_get_node_type_count() const;
+	Dictionary _b_get_node_type_info(int type_id) const;
+	Array _b_get_connections() const;
+	// TODO Only exists because the UndoRedo API is confusing `null` with `absence of argument`...
+	// See https://github.com/godotengine/godot/issues/36895
+	void _b_set_node_param_null(int node_id, int param_index);
+	void _b_set_node_name(int node_id, String name);
+
+	static void _bind_methods();
+
+	ProgramGraph _graph;
+};
+
+ProgramGraph::Node *create_node_internal(ProgramGraph &graph, VoxelGraphFunction::NodeTypeID type_id, Vector2 position,
+		uint32_t id, bool create_default_instances);
+
+} // namespace zylann::voxel
+
+ZN_GODOT_VARIANT_ENUM_CAST(zylann::voxel::VoxelGraphFunction, NodeTypeID)
+
+#endif // VOXEL_GRAPH_FUNCTION_H
